move useTheme hook to the context
