--- src/hooks/index.js
@@ -1,1 +1,0 @@
-export * from './useTheme'
